requirements.txt
yt-dlp
openai-whisper
ultralytics
opencv-python-headless
pytesseract
ffmpeg-python
torch
torchvision
torchaudio
